requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
cloudscraper==1.2.40
elasticsearch==7.7.1
feedparser==5.2.1
idna==2.9
lxml==4.6.2
pika==1.1.0
pymongo==3.10.1
pyparsing==2.4.7
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
requests-toolbelt==0.9.1
schedule==0.6.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.9
